PANA-7681 Extract image snapshot test mocks
